requirements.txt
absl-py==2.1.0
accelerate==1.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
attrs==24.3.0
beautifulsoup4==4.12.3
bitsandbytes==0.45.0
certifi==2024.12.14
charset-normalizer==3.4.0
colorama==0.4.4
colorlog==4.7.2
datasets==3.2.0
dill==0.3.8
einops==0.8.0
filelock==3.13.1
frozenlist==1.5.0
fsspec==2024.2.0
gdown==5.2.0
grpcio==1.68.1
huggingface-hub==0.27.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
kmeans-pytorch==0.3
Markdown==3.7
MarkupSafe==2.1.5
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
networkx==3.2.1
nltk==3.9.1
numpy==1.26.3
openai==1.76.0
packaging==24.2
pandas==2.2.3
peft==0.14.0
pillow==10.2.0
plotly==5.24.1
propcache==0.2.1
protobuf==5.29.2
psutil==6.1.0
pyarrow==18.1.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
recbole==1.2.0
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.6.0
scipy==1.14.1
sentence-transformers==3.3.1
setuptools==75.1.0
six==1.17.0
soupsieve==2.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
texttable==1.7.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tiktoken==0.9.0
tokenizers==0.21.0
torch==2.5.0 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.5.0 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.20.0 --index-url https://download.pytorch.org/whl/cu121
tqdm==4.67.1
transformers==4.47.1
triton==3.1.0
typing_extensions==4.9.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wheel==0.44.0
xxhash==3.5.0
yarl==1.18.3
